feat(ui): better layout for generator load from file buttons
