+++ invokeai/frontend/web/src/features/nodes/components/flow/nodes/Invocation/fields/inputs/StringGeneratorDynamicPromptsCombinatorialSettings.tsx
@@ -1,7 +1,6 @@
-import { CompositeNumberInput, Flex, FormControl, FormLabel, Textarea } from '@invoke-ai/ui-library';
-import { LoadTextFromFileIconButton } from 'features/nodes/components/flow/nodes/Invocation/fields/inputs/LoadTextFromFileIconButton';
+import { CompositeNumberInput, Flex, FormControl, FormLabel } from '@invoke-ai/ui-library';
+import { GeneratorTextareaWithFileUpload } from 'features/nodes/components/flow/nodes/Invocation/fields/inputs/GeneratorTextareaWithFileUpload';
 import type { StringGeneratorDynamicPromptsCombinatorial } from 'features/nodes/types/field';
-import type { ChangeEvent } from 'react';
 import { memo, useCallback, useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDynamicPromptsQuery } from 'services/api/endpoints/utilities';
@@ -17,8 +16,8 @@ export const StringGeneratorDynamicPromptsCombinatorialSettings = memo(
     const loadingValues = useMemo(() => [`<${t('nodes.generatorLoading')}>`], [t]);
 
     const onChangeInput = useCallback(
-      (e: ChangeEvent<HTMLTextAreaElement>) => {
-        onChange({ ...state, input: e.target.value, values: loadingValues });
+      (input: string) => {
+        onChange({ ...state, input, values: loadingValues });
       },
       [onChange, state, loadingValues]
     );
@@ -47,32 +46,13 @@ export const StringGeneratorDynamicPromptsCombinatorialSettings = memo(
       }
     }, [data, isLoading, loadingValues, onChange, state]);
 
-    const onLoadFile = useCallback(
-      (value: string) => {
-        onChange({ ...state, input: value });
-      },
-      [onChange, state]
-    );
-
     return (
       <Flex gap={2} flexDir="column">
         <FormControl orientation="vertical">
           <FormLabel>{t('dynamicPrompts.maxPrompts')}</FormLabel>
           <CompositeNumberInput value={state.maxPrompts} onChange={onChangeMaxPrompts} min={1} max={1000} w="full" />
         </FormControl>
-        <FormControl orientation="vertical" position="relative">
-          <FormLabel>{t('common.input')}</FormLabel>
-          <Textarea
-            className="nowheel nodrag nopan"
-            value={state.input}
-            onChange={onChangeInput}
-            p={2}
-            resize="none"
-            rows={5}
-            fontSize="sm"
-          />
-          <LoadTextFromFileIconButton position="absolute" top={10} right={2} onLoadFile={onLoadFile} />
-        </FormControl>
+        <GeneratorTextareaWithFileUpload value={state.input} onChange={onChangeInput} />
       </Flex>
     );
   }
