+++ invokeai/frontend/web/src/features/nodes/components/flow/nodes/Invocation/fields/inputs/StringGeneratorParseStringSettings.tsx
@@ -1,5 +1,5 @@
-import { Flex, FormControl, FormLabel, Input, Textarea } from '@invoke-ai/ui-library';
-import { LoadTextFromFileIconButton } from 'features/nodes/components/flow/nodes/Invocation/fields/inputs/LoadTextFromFileIconButton';
+import { Flex, FormControl, FormLabel, Input } from '@invoke-ai/ui-library';
+import { GeneratorTextareaWithFileUpload } from 'features/nodes/components/flow/nodes/Invocation/fields/inputs/GeneratorTextareaWithFileUpload';
 import type { StringGeneratorParseString } from 'features/nodes/types/field';
 import type { ChangeEvent } from 'react';
 import { memo, useCallback } from 'react';
@@ -21,15 +21,8 @@ export const StringGeneratorParseStringSettings = memo(
     );
 
     const onChangeInput = useCallback(
-      (e: ChangeEvent<HTMLTextAreaElement>) => {
-        onChange({ ...state, input: e.target.value });
-      },
-      [onChange, state]
-    );
-
-    const onLoadFile = useCallback(
-      (value: string) => {
-        onChange({ ...state, input: value });
+      (input: string) => {
+        onChange({ ...state, input });
       },
       [onChange, state]
     );
@@ -40,19 +33,7 @@ export const StringGeneratorParseStringSettings = memo(
           <FormLabel>{t('nodes.splitOn')}</FormLabel>
           <Input value={state.splitOn} onChange={onChangeSplitOn} />
         </FormControl>
-        <FormControl position="relative" orientation="vertical">
-          <FormLabel>{t('common.input')}</FormLabel>
-          <Textarea
-            className="nowheel nodrag nopan"
-            value={state.input}
-            onChange={onChangeInput}
-            p={2}
-            resize="none"
-            rows={5}
-            fontSize="sm"
-          />
-          <LoadTextFromFileIconButton position="absolute" top={10} right={2} onLoadFile={onLoadFile} />
-        </FormControl>
+        <GeneratorTextareaWithFileUpload value={state.input} onChange={onChangeInput} />
       </Flex>
     );
   }
